pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "test task"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.124.0",
    "httpx>=0.27.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "asyncpg>=0.30.0",
    "uvicorn>=0.38.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "gspread>=6.2.1",
    "google-auth>=2.43.0",
    "numpy>=2.3.5",
    "aiohttp>=3.13.2",
    "ollama>=0.6.1",
    "rapidfuzz>=3.14.3",
    "aiogram>=3.13.1",
    "aiosqlite>=0.20.0",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]

[tool.ruff]
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused import
"tests/*" = ["S101"]  # use of assert in tests

[tool.ruff.lint.isort]
known-first-party = ["my_package"]
